requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.0
nltk==3.5
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.15.0
regex==2020.11.13
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
textblob==0.15.3
tqdm==4.55.1
tweepy==3.10.0
urllib3==1.26.2
